Create README.md
GoodWood is a automated system designed for the control of the wood flow around colombia, although it can be applied to another countries with similar conditions.

Goodwood is a solution based on the consumer, our philosophy is that the final consumer is able to cause impact in the decision of the big companies that sells wood. Our team have proposed a logo that will be given only to the companies that are commited with the responsable exploitation of the raw material. But how can we assure that?

Goodwood is a Hardware and software solution, that checks the state of the wood carried by a transporter, our system measures the position and the weight of the truck that carries a certain type of wood. Every time that the systems detects an anomaly in the system it saves the position and the weight of the truck in that specific moment. This information will be sent to a server when it pass in the control borders stablished in colombia, and our system will compare the data to the permissions given by the authorities of the contry. If the loaded place or the amount of wood is irregular with the permission, then goodwood will inform the authorities about the irregularity. 


This will help us to assume control and to be able to give the logo only to the companies that are not committing irregularities in the process of the wood transportation, given the opportunity to the consumer to choose products that are committed with the environment . 



In order to run this this project it's necessary to change the IP address of the components to match the server (Python Flask file) address.